printf basic implementation is done, error handling and u-specifier(at neg number fails) are done partially

+++ hexadecimal.c
@@ -12,7 +12,7 @@
 
 #include "libftprintf.h"
 
-static int	get_size(unsigned int integer)
+static int	get_size(unsigned long integer)
 {
 	if (integer < 16 && integer >= 0) // (integer < 16 && integer >= -16)
 		return (1);
@@ -38,7 +38,7 @@ static char	hexadecimal_value(int index, char letter_size)
 	}
 }
 
-static char	*hexadecimal_positive(unsigned int integer, char letter_size,
+static char	*hexadecimal_positive(unsigned long integer, char letter_size,
 		char *hexa)
 {
 	if (integer >= 16)
@@ -63,14 +63,14 @@ static char	*hexadecimal_positive(unsigned int integer, char letter_size,
 // 	return (pos + 1);
 // }
 
-char	*decimal_to_hexadecimal(int integer, char letter_size)
+char	*decimal_to_hexadecimal(long integer, char letter_size)
 {
 	char			*hexa;
 	char			*end;
-	unsigned int	num;
+	unsigned long	num;
 
 	// int				pos;
-	num = (unsigned int)integer;
+	num = (unsigned long)integer;
 	hexa = malloc(sizeof(char) * (get_size(num) + 1));
 	end = hexadecimal_positive(num, letter_size, hexa);
 	// printf("%d == ", get_size(integer));
